api_requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
openai==1.6.1
azure-search-documents==11.5.3
python-dotenv==1.0.0
